task1

+++ 0x0B-malloc_free/1-strdup.c
@@ -9,9 +9,9 @@
 
 char *_strdup(char *str)
 {
-	size_t g;
+	int g;
 	char *new_st;
-	size_t x;
+	int x;
 
 	if (str == NULL)
 	{
